# MergeIterator: allocate less memory at first