--- a/pkg/querier/batch/chunk_test.go
+++ b/pkg/querier/batch/chunk_test.go
@@ -59,6 +59,7 @@ func mkChunk(t require.TestingT, from model.Time, points int, enc promchunk.Enco
 		require.Nil(t, npc)
 		ts = ts.Add(step)
 	}
+	ts = ts.Add(-step) // undo the add that we did just before exiting the loop
 	return chunk.NewChunk(userID, fp, metric, pc, from, ts)
 }